UI touchups and debug panel
